+++ crates/nostr/src/lib.rs
@@ -12,7 +12,7 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 #![cfg_attr(bench, feature(test))]
 #![cfg_attr(docsrs, feature(doc_auto_cfg))]
-#![cfg_attr(feature = "default", doc = include_str!("../README.md"))]
+#![cfg_attr(all(feature = "std", feature = "all-nips"), doc = include_str!("../README.md"))]
 
 #[cfg(not(any(feature = "std", feature = "alloc")))]
 compile_error!("at least one of the `std` or `alloc` features must be enabled");