+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 
 ### Changed
 
+* nostr: disable all default features except `std` ([Yuki Kishimoto])
 * pool: rework latency tracking ([Yuki Kishimoto])
 * pool: switch from async to sync message sending for `Relay` ([Yuki Kishimoto])